requirements.txt
pandas
google-generativeai
rich
python-dotenv
plotext
